This PR introduces a few changes:

* Migrate from `ds.Batching` to `ds.TxnDatastore`.
* Optimisations in overall logic to improved speed and allocations.
* Cache logic changes: it is now a read-through cache, and we store `peer.ID` => `[]ma.Multiaddr` entries. 
  * The role of the cache is to avoid disk IO for popular entries.
  * We now have the guarantee that if a peer is present, the list of multiaddrs is complete and none of them are expired.
  * We only do blanket operations now. Cache entries are only loaded on cache misses; write operations simply evict the cache. No detailed bookkeeping, to keep cache maintenance overhead low.
  * We can now serve `Addrs()` calls fully from cache without hitting the datastore, or performing range scans.
* Configuration struct for DS Peerstore.
* Ability to turn off the cache (esp. for benchmarks).
* Parameterised benchmarks, and enhanced coverage.

## Benchmarks

We now benchmark these operations separately:

* AddAddrs => measures adding an address (random one each time).
* SetAddrs => measures setting an address (random one each time).
* GetAddrs => measures getting an already set address (same one every time).
* AddGetAndClearAddrs => adds, gets and clears an address, all within the benchmark.
* Get1000PeersWithAddrs => adds 1000 peers outside the benchmark, and calls `PeersWithAddrs()`.

Each benchmark is run with peers containing 1, 10 and 100 multiaddrs. 

Additionally, datastore peerstore benchmarks run in two variants: with and without the cache. This allows us to characterise performance under a cache-favorable usage scenario and cache-unfriendly one.

## benchcmp

between dcb3a93288741c8d9377cd270c76729b5072f98d and this PR.

Big improvements in GetAddrs thanks to better cache utilisation, end-to-end cases (AddGetClear), and generally as records scale beyond 1 multiaddr.

```
→ benchcmp bench{1,2}.txt
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkBadgerDsPeerstore/AddAddrs-1Addrs-Caching-8                    123160        110075        -10.62%
BenchmarkBadgerDsPeerstore/AddAddrs-10Addrs-Caching-8                   294927        246947        -16.27%
BenchmarkBadgerDsPeerstore/AddAddrs-100Addrs-Caching-8                  1195351       722421        -39.56%
BenchmarkBadgerDsPeerstore/SetAddrs-1Addrs-Caching-8                    117037        106698        -8.83%
BenchmarkBadgerDsPeerstore/SetAddrs-10Addrs-Caching-8                   285869        247638        -13.37%
BenchmarkBadgerDsPeerstore/SetAddrs-100Addrs-Caching-8                  1220655       734192        -39.85%
BenchmarkBadgerDsPeerstore/GetAddrs-1Addrs-Caching-8                    22488         2927          -86.98%
BenchmarkBadgerDsPeerstore/GetAddrs-10Addrs-Caching-8                   34173         2990          -91.25%
BenchmarkBadgerDsPeerstore/GetAddrs-100Addrs-Caching-8                  106863        3391          -96.83%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-1Addrs-Caching-8         1258245       831498        -33.92%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-10Addrs-Caching-8        6061845       3532039       -41.73%
BenchmarkBadgerDsPeerstore/AddGetAndClearAddrs-100Addrs-Caching-8       13373294      9780208       -26.87%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-1Addrs-Caching-8       5968105       5142968       -13.83%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-10Addrs-Caching-8      39170800      15715752      -59.88%
BenchmarkBadgerDsPeerstore/Get1000PeersWithAddrs-100Addrs-Caching-8     390117659     127612668     -67.29%
```

Outputs in Gist: https://gist.github.com/raulk/e5cbf2b3037c86df2461d3c159f13b2a