@raulk 2018-09-11
@bigs @lanzafame @Stebalien this PR is ready for review. Up next:

* DB-managed TTL.
* quick migration of `KeyBook` and `PeerMetadata` to datastore (not sure if it's needed, but it's odd to have half and half).
* explore more efficient key tuple designs.