@raulk 2018-09-11
Before the merge I plan on adding two more tests to increase coverage:

* cacheless addressbook (it is only tested in benchmarks now, so coverage doesn't capture it)
* `dbDeleteIter()`